🚚 Rename: activity-form 폴더 구조를 목적에 따라 'common', 'new', 'edit'으로 분리
