+++ src/features/my-page/activity-form/common/constants/category.ts
