+++ src/features/my-page/activity-form/common/ui/image-uploader/DeleteButton.tsx
