+++ src/features/my-page/activity-form/common/ui/schedule-selector/schedule-calendar.css
