+++ src/features/my-page/activity-form/common/utils/images.ts
